feat: Implement Phase 3 precision and indexing verification for PyTorch implementation
This commit addresses the following key improvements:
- Ensures structure_factors_batch returns complex128 dtype
- Adds dtype assertions in apply_disorder method
- Implements comprehensive indexing tests for grid and Brillouin zone conversions
- Updates Winv comparison in test_compute_gnm_phonons_equivalence to handle NaN values more robustly

The changes include:
- Explicit dtype casting in scatter_torch.py
- Dtype assertions in models_torch.py
- New test cases in test_models_torch_kvector.py for indexing functions
- Updated NaN handling in test_models_torch_phonon.py

These modifications improve numerical precision and indexing correctness in the PyTorch implementation.

+++ eryx/scatter_torch.py
@@ -182,8 +182,12 @@ def structure_factors_batch(q_grid: torch.Tensor, xyz: torch.Tensor,
         A_real = torch.sum(A_real, dim=1)
         A_imag = torch.sum(A_imag, dim=1)
     
+    # Ensure A_real and A_imag are float64 before creating complex tensor
+    A_real = A_real.to(dtype=torch.float64)
+    A_imag = A_imag.to(dtype=torch.float64)
+    
     # Return complex structure factors with high precision
-    return torch.complex(A_real, A_imag).to(dtype=torch.complex128)
+    return torch.complex(A_real, A_imag)
 
 def structure_factors(q_grid: torch.Tensor, xyz: torch.Tensor, 
                      ff_a: torch.Tensor, ff_b: torch.Tensor, ff_c: torch.Tensor,
